First Draft
The scorer does not yet work, issues because the word set needs to be created before the function even though its created in the beginning of the function. 
Remover is very simple and not efficient, but quick anyways because the dictionary is not that large, so might be fine to keep it like this. 
The Main is not good at all yet, needs most changes. Once I figure the scorer out I will create a solid main document that makes use of the scorer and remover and simplifies their use